Get raw data

A function, that returns all, unprocessed data concerning one of the offices in Warsaw.

get_raw_data(office_name)

Arguments

office_name
character acronym of office in Warsaw.

You can get a list of possible values using get_available_offices function.

Value

A data.frame with following columns:

status - either 0 (queue is not operating) or 1 (queue is operating).

czasObslugi - expected time of waiting in queue, in minutes. See also: get_waiting_time.

lp - ordinal number.

idGrupy - ID of a queue from nazwaGrupy.

liczbaCzynnychStan - amount of opened counters. See also: get_open_counters.

nazwaGrupy - a name of a queue. See also: get_available_queues.

literaGrupy - a single letter symbolizing a queue name from nazwaGrupy.

liczbaKlwKolejce - amount of people in queue. See also: get_number_of_people.

aktualnyNumer - current ticket number. See also: get_current_ticket_number.

See also

Examples

office <- get_available_offices()[1]
get_raw_data(office)

#>    status czasObslugi lp idGrupy liczbaCzynnychStan
#> 1       0          10  1     135                  0
#> 2       0          10  2     127                  0
#> 3       0          10  3     119                  0
#> 4       0          10  4     128                  0
#> 5       0          10  5     118                  0
#> 6       0          10  6     130                  0
#> 7       0          10  7     126                  0
#> 8       0          10  8     224                  0
#> 9       0          10  9     120                  0
#> 10      0          10 10     129                  0
#> 11      0          10 11     122                  0
#> 12      0          10 12      94                  0
#> 13      0          10 13     131                  0
#> 14      0          10 14     132                  0
#> 15      0          10 15     115                  0
#> 16      0          60 16     133                  0
#> 17      0          60 17     222                  0
#> 18      0          10 18     121                  0
#> 19      0          10 19     134                  0
#>                                                       nazwaGrupy literaGrupy
#> 1              T: Odbiór odpisu po sprostowaniu lub uzupełnieniu           T
#> 2                   Z: Zgłoszenie zgonu na podstawie karty zgonu           Z
#> 3  P: Oświadczenie o powrocie do nazwiska noszonego przed ślubem           P
#> 4                          S: Sprostowanie lub uzupełnienie aktu           S
#> 5       I: Zmiana imienia lub nazwiska w trybie administracyjnym           I
#> 6                    Z: Sprostowanie lub uzupełnienie aktu zgonu           Z
#> 7                  J: Zgłoszenie jubileuszu pożycia małżeńskiego           J
#> 8                            Sprostowanie lub uzupełnienie aktów           S
#> 9                                      D: Zmiana imienia dziecka           D
#> 10              M: Sprostowanie lub uzupełnienie aktu małżeństwa           M
#> 11                M: Złożenie dokumentów i wydawanie zaświadczeń           M
#> 12                                     A: Rejestracja noworodków           N
#> 13          T: Rejestracja na podstawie zagranicznych dokumentów           T
#> 14     T: Odbiór odpisu po rejestracji  zagranicznych dokumentów           T
#> 15           W: Złożenie wniosku oraz odbiór zamówionych odpisów           W
#> 16                                                       X: Kasa           X
#> 17                    C: Wydanie zaświadczenia o stanie cywilnym           C
#> 18              O: Uznanie ojcostwa przed urodzeniem się dziecka           O
#> 19                         T: Sprostowanie lub uzupełnienie aktu           T
#>    liczbaKlwKolejce aktualnyNumer
#> 1                 0              
#> 2                 0              
#> 3                 0              
#> 4                 0              
#> 5                 0              
#> 6                 0              
#> 7                 0              
#> 8                 0              
#> 9                 0              
#> 10                0              
#> 11                0              
#> 12                0              
#> 13                0              
#> 14                0              
#> 15                0              
#> 16                0              
#> 17                0              
#> 18                0              
#> 19                0