Get specific data

Several functions to get specific data, such as waiting time, open encounters, current ticket number nad amount of people in a specific queue in specified office.

get_waiting_time(office_name, queue_name)

get_open_counters(office_name, queue_name)

get_current_ticket_number(office_name, queue_name)

get_number_of_people(office_name, queue_name)

Arguments

office_name
character acronym of office in Warsaw.

You can get a list of possible values using get_available_offices function.

queue_name
A character describing a queue we are interested in.

You can get a list of possible values using get_available_queues function.

Value

A character in format depending on the called function.

If get_waiting_time is called:

"Waiting time for <queue name> is x minutes".

If get_open_encounters is called:

"There are x open encounters for <queue name>".

If get_current_number is called:

"Current ticket number is x"

If get_number_of_people is called:

"There are x people in <queue name>"

Functions

get_waiting_time: Returns expected time to be served.

get_open_counters: Returns amount of opened encounters.

get_current_ticket_number: Returns current ticket number.

get_number_of_people: Returns number of people waiting in specified queue.

Examples

office <- get_available_offices()[1]
queue <- get_available_queues(office)

get_waiting_time(office, queue)

#>  [1] "Waiting time for T: Odbiór odpisu po sprostowaniu lub uzupełnieniu is 10 minutes."            
#>  [2] "Waiting time for Z: Zgłoszenie zgonu na podstawie karty zgonu is 10 minutes."                 
#>  [3] "Waiting time for P: Oświadczenie o powrocie do nazwiska noszonego przed ślubem is 10 minutes."
#>  [4] "Waiting time for S: Sprostowanie lub uzupełnienie aktu is 10 minutes."                        
#>  [5] "Waiting time for I: Zmiana imienia lub nazwiska w trybie administracyjnym is 10 minutes."     
#>  [6] "Waiting time for Z: Sprostowanie lub uzupełnienie aktu zgonu is 10 minutes."                  
#>  [7] "Waiting time for J: Zgłoszenie jubileuszu pożycia małżeńskiego is 10 minutes."                
#>  [8] "Waiting time for Sprostowanie lub uzupełnienie aktów is 10 minutes."                          
#>  [9] "Waiting time for D: Zmiana imienia dziecka is 10 minutes."                                    
#> [10] "Waiting time for M: Sprostowanie lub uzupełnienie aktu małżeństwa is 10 minutes."             
#> [11] "Waiting time for M: Złożenie dokumentów i wydawanie zaświadczeń is 10 minutes."               
#> [12] "Waiting time for A: Rejestracja noworodków is 10 minutes."                                    
#> [13] "Waiting time for T: Rejestracja na podstawie zagranicznych dokumentów is 10 minutes."         
#> [14] "Waiting time for T: Odbiór odpisu po rejestracji  zagranicznych dokumentów is 10 minutes."    
#> [15] "Waiting time for W: Złożenie wniosku oraz odbiór zamówionych odpisów is 10 minutes."          
#> [16] "Waiting time for X: Kasa is 60 minutes."                                                      
#> [17] "Waiting time for C: Wydanie zaświadczenia o stanie cywilnym is 60 minutes."                   
#> [18] "Waiting time for O: Uznanie ojcostwa przed urodzeniem się dziecka is 10 minutes."             
#> [19] "Waiting time for T: Sprostowanie lub uzupełnienie aktu is 10 minutes."                        

get_open_counters(office, queue)

#>  [1] "There are  0 open counters for  T: Odbiór odpisu po sprostowaniu lub uzupełnieniu"            
#>  [2] "There are  0 open counters for  Z: Zgłoszenie zgonu na podstawie karty zgonu"                 
#>  [3] "There are  0 open counters for  P: Oświadczenie o powrocie do nazwiska noszonego przed ślubem"
#>  [4] "There are  0 open counters for  S: Sprostowanie lub uzupełnienie aktu"                        
#>  [5] "There are  0 open counters for  I: Zmiana imienia lub nazwiska w trybie administracyjnym"     
#>  [6] "There are  0 open counters for  Z: Sprostowanie lub uzupełnienie aktu zgonu"                  
#>  [7] "There are  0 open counters for  J: Zgłoszenie jubileuszu pożycia małżeńskiego"                
#>  [8] "There are  0 open counters for  Sprostowanie lub uzupełnienie aktów"                          
#>  [9] "There are  0 open counters for  D: Zmiana imienia dziecka"                                    
#> [10] "There are  0 open counters for  M: Sprostowanie lub uzupełnienie aktu małżeństwa"             
#> [11] "There are  0 open counters for  M: Złożenie dokumentów i wydawanie zaświadczeń"               
#> [12] "There are  0 open counters for  A: Rejestracja noworodków"                                    
#> [13] "There are  0 open counters for  T: Rejestracja na podstawie zagranicznych dokumentów"         
#> [14] "There are  0 open counters for  T: Odbiór odpisu po rejestracji  zagranicznych dokumentów"    
#> [15] "There are  0 open counters for  W: Złożenie wniosku oraz odbiór zamówionych odpisów"          
#> [16] "There are  0 open counters for  X: Kasa"                                                      
#> [17] "There are  0 open counters for  C: Wydanie zaświadczenia o stanie cywilnym"                   
#> [18] "There are  0 open counters for  O: Uznanie ojcostwa przed urodzeniem się dziecka"             
#> [19] "There are  0 open counters for  T: Sprostowanie lub uzupełnienie aktu"                        

get_current_ticket_number(office, queue)

#>  [1] "Current ticket number is  "     "Current ticket number is  Z603"
#>  [3] "Current ticket number is  P601" "Current ticket number is  "    
#>  [5] "Current ticket number is  I016" "Current ticket number is  "    
#>  [7] "Current ticket number is  "     "Current ticket number is  "    
#>  [9] "Current ticket number is  "     "Current ticket number is  "    
#> [11] "Current ticket number is  M015" "Current ticket number is  "    
#> [13] "Current ticket number is  "     "Current ticket number is  "    
#> [15] "Current ticket number is  W091" "Current ticket number is  "    
#> [17] "Current ticket number is  "     "Current ticket number is  "    
#> [19] "Current ticket number is  "    

get_number_of_people(office, queue)

#>  [1] "There are  0  people in  T: Odbiór odpisu po sprostowaniu lub uzupełnieniu"            
#>  [2] "There are  0  people in  Z: Zgłoszenie zgonu na podstawie karty zgonu"                 
#>  [3] "There are  0  people in  P: Oświadczenie o powrocie do nazwiska noszonego przed ślubem"
#>  [4] "There are  0  people in  S: Sprostowanie lub uzupełnienie aktu"                        
#>  [5] "There are  0  people in  I: Zmiana imienia lub nazwiska w trybie administracyjnym"     
#>  [6] "There are  0  people in  Z: Sprostowanie lub uzupełnienie aktu zgonu"                  
#>  [7] "There are  0  people in  J: Zgłoszenie jubileuszu pożycia małżeńskiego"                
#>  [8] "There are  0  people in  Sprostowanie lub uzupełnienie aktów"                          
#>  [9] "There are  0  people in  D: Zmiana imienia dziecka"                                    
#> [10] "There are  0  people in  M: Sprostowanie lub uzupełnienie aktu małżeństwa"             
#> [11] "There are  0  people in  M: Złożenie dokumentów i wydawanie zaświadczeń"               
#> [12] "There are  0  people in  A: Rejestracja noworodków"                                    
#> [13] "There are  0  people in  T: Rejestracja na podstawie zagranicznych dokumentów"         
#> [14] "There are  0  people in  T: Odbiór odpisu po rejestracji  zagranicznych dokumentów"    
#> [15] "There are  0  people in  W: Złożenie wniosku oraz odbiór zamówionych odpisów"          
#> [16] "There are  0  people in  X: Kasa"                                                      
#> [17] "There are  0  people in  C: Wydanie zaświadczenia o stanie cywilnym"                   
#> [18] "There are  1  people in  O: Uznanie ojcostwa przed urodzeniem się dziecka"             
#> [19] "There are  0  people in  T: Sprostowanie lub uzupełnienie aktu"