Get specific data verbosely

Several functions to get specific data, such as waiting time, open encounters, current ticket number and amount of people in a specific queue in specified office.

get_waiting_time_verbose(office_name, queue_name, language = "en")

get_open_counters_verbose(office_name, queue_name, language = "en")

get_current_ticket_number_verbose(office_name, queue_name,
  language = "en")

get_number_of_people_verbose(office_name, queue_name, language = "en")

Arguments

office_name
character acronym of office in Warsaw.

You can get a list of possible values using get_available_offices function.

queue_name
A character describing a queue we are interested in. You can get a list of possible values using get_available_queues function.

language
A character. Only two languages supported: english ("en") and polish ("pl").

Value

A character in format depending on the called function and the variable language. Below we assume, that language variable is default.

If get_waiting_time_verbose is called:

"Waiting time for <queue name> is x minutes".

If get_open_encounters_verbose is called:

"There are x open encounters for <queue name>".

If get_current_number_verbose is called:

"Current ticket number is x"

If get_number_of_people_verbose is called:

"There are x people in <queue name>"

Functions

get_waiting_time_verbose: Returns expected time to be served.

get_open_counters_verbose: Returns amount of opened encounters.

get_current_ticket_number_verbose: Returns current ticket number.

get_number_of_people_verbose: Returns number of people waiting in specified queue.

See also

Examples

office <- get_available_offices()[1]
queue <- get_available_queues(office)

get_waiting_time_verbose(office, queue)

#> character(0)

get_open_counters_verbose(office, queue)

#>  [1] "There are 0 open counters for T: Odbiór odpisu po sprostowaniu lub uzupełnieniu"            
#>  [2] "There are 0 open counters for Z: Zgłoszenie zgonu na podstawie karty zgonu"                 
#>  [3] "There are 0 open counters for P: Oświadczenie o powrocie do nazwiska noszonego przed ślubem"
#>  [4] "There are 0 open counters for S: Sprostowanie lub uzupełnienie aktu"                        
#>  [5] "There are 0 open counters for I: Zmiana imienia lub nazwiska w trybie administracyjnym"     
#>  [6] "There are 0 open counters for Z: Sprostowanie lub uzupełnienie aktu zgonu"                  
#>  [7] "There are 0 open counters for J: Zgłoszenie jubileuszu pożycia małżeńskiego"                
#>  [8] "There are 0 open counters for Sprostowanie lub uzupełnienie aktów"                          
#>  [9] "There are 0 open counters for D: Zmiana imienia dziecka"                                    
#> [10] "There are 0 open counters for M: Sprostowanie lub uzupełnienie aktu małżeństwa"             
#> [11] "There are 0 open counters for M: Złożenie dokumentów i wydawanie zaświadczeń"               
#> [12] "There are 0 open counters for A: Rejestracja noworodków"                                    
#> [13] "There are 0 open counters for T: Rejestracja na podstawie zagranicznych dokumentów"         
#> [14] "There are 0 open counters for T: Odbiór odpisu po rejestracji  zagranicznych dokumentów"    
#> [15] "There are 0 open counters for W: Złożenie wniosku oraz odbiór zamówionych odpisów"          
#> [16] "There are 0 open counters for X: Kasa"                                                      
#> [17] "There are 0 open counters for C: Wydanie zaświadczenia o stanie cywilnym"                   
#> [18] "There are 0 open counters for O: Uznanie ojcostwa przed urodzeniem się dziecka"             
#> [19] "There are 0 open counters for T: Sprostowanie lub uzupełnienie aktu"                        

get_current_ticket_number_verbose(office, queue)

#> Warning: the condition has length > 1 and only the first element will be used

#> [1] "Obecny numerek w kolejce to: 0."

get_number_of_people_verbose(office, queue)

#>  [1] "There are 0 people in T: Odbiór odpisu po sprostowaniu lub uzupełnieniu."            
#>  [2] "There are 0 people in Z: Zgłoszenie zgonu na podstawie karty zgonu."                 
#>  [3] "There are 0 people in P: Oświadczenie o powrocie do nazwiska noszonego przed ślubem."
#>  [4] "There are 0 people in S: Sprostowanie lub uzupełnienie aktu."                        
#>  [5] "There are 0 people in I: Zmiana imienia lub nazwiska w trybie administracyjnym."     
#>  [6] "There are 0 people in Z: Sprostowanie lub uzupełnienie aktu zgonu."                  
#>  [7] "There are 0 people in J: Zgłoszenie jubileuszu pożycia małżeńskiego."                
#>  [8] "There are 0 people in Sprostowanie lub uzupełnienie aktów."                          
#>  [9] "There are 0 people in D: Zmiana imienia dziecka."                                    
#> [10] "There are 0 people in M: Sprostowanie lub uzupełnienie aktu małżeństwa."             
#> [11] "There are 0 people in M: Złożenie dokumentów i wydawanie zaświadczeń."               
#> [12] "There are 0 people in A: Rejestracja noworodków."                                    
#> [13] "There are 0 people in T: Rejestracja na podstawie zagranicznych dokumentów."         
#> [14] "There are 0 people in T: Odbiór odpisu po rejestracji  zagranicznych dokumentów."    
#> [15] "There are 0 people in W: Złożenie wniosku oraz odbiór zamówionych odpisów."          
#> [16] "There are 0 people in X: Kasa."                                                      
#> [17] "There are 0 people in C: Wydanie zaświadczenia o stanie cywilnym."                   
#> [18] "There are 0 people in O: Uznanie ojcostwa przed urodzeniem się dziecka."             
#> [19] "There are 0 people in T: Sprostowanie lub uzupełnienie aktu."